Move version command to internal package
Keeping it in pkg could lead to confusion since it only gets
populated when the receptor binary is build with the Makefile,
not when importing it.
